Refactored data connection code into separate file
