+++ amp/app/data-integration/capturejs-disable.js
