+++ amp/app/data-integration/constants.js
@@ -1,0 +1,2 @@
+export const PAGE_TITLE = 'pageTitle'
+export const DATA_INIT_FUNCTION = 'dataInitFunction'
